Break out responsibilities of LobbyLoginValidator to new helper classes
