+++ lobby/src/main/java/org/triplea/lobby/server/login/LobbyLoginValidator.java
@@ -1,19 +1,14 @@
 package org.triplea.lobby.server.login;
 
 import com.google.common.annotations.VisibleForTesting;
-import games.strategy.engine.lobby.PlayerEmailValidation;
 import games.strategy.engine.lobby.PlayerNameValidation;
 import games.strategy.net.ILoginValidator;
-import games.strategy.net.MacFinder;
 import games.strategy.net.nio.ServerQuarantineConversation;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
-import java.sql.Timestamp;
-import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
 import lombok.AllArgsConstructor;
@@ -26,7 +21,6 @@ import org.triplea.lobby.server.User;
 import org.triplea.lobby.server.db.DatabaseDao;
 import org.triplea.lobby.server.db.HashedPassword;
 import org.triplea.lobby.server.login.forgot.password.verify.TempPasswordVerification;
-import org.triplea.util.Version;
 
 /**
  * The server side of the lobby authentication protocol.
@@ -42,10 +36,6 @@ public final class LobbyLoginValidator implements ILoginValidator {
   interface ErrorMessages {
     String ANONYMOUS_AUTHENTICATION_FAILED = "Can't login anonymously, username already exists";
     String AUTHENTICATION_FAILED = "Incorrect username or password";
-    String INVALID_MAC = "Invalid mac address";
-    String THAT_IS_NOT_A_NICE_NAME = "That's not a nice name";
-    String USERNAME_HAS_BEEN_BANNED = "This username is banned, please create a new one.";
-    String YOU_HAVE_BEEN_BANNED = "You have been banned from the TripleA lobby.";
     String TOO_MANY_FAILED_LOGIN_ATTEMPTS =
         "Too many failed login attempts, wait a few minutes before attempting again.";
   }
@@ -55,6 +45,8 @@ public final class LobbyLoginValidator implements ILoginValidator {
   private final Supplier<String> bcryptSaltGenerator;
   private final FailedLoginThrottle failedLoginThrottle;
   private final TempPasswordVerification tempPasswordVerification;
+  private final AllowLoginRules allowLoginRules;
+  private final AllowCreateUserRules allowCreateUserRules;
 
   @Override
   public Map<String, String> getChallengeProperties(final String username) {
@@ -72,7 +64,7 @@ public final class LobbyLoginValidator implements ILoginValidator {
 
     final var address = ((InetSocketAddress) remoteAddress).getAddress();
     if (failedLoginThrottle.tooManyFailedLoginAttempts(address)) {
-      return ErrorMessages.TOO_MANY_FAILED_LOGIN_ATTEMPTS;
+      return LobbyLoginValidator.ErrorMessages.TOO_MANY_FAILED_LOGIN_ATTEMPTS;
     }
 
     final User user =
@@ -81,7 +73,10 @@ public final class LobbyLoginValidator implements ILoginValidator {
             .inetAddress(((InetSocketAddress) remoteAddress).getAddress())
             .hashedMacAddress(clientMac)
             .build();
-    final @Nullable String errorMessage = authenticateUser(response, user);
+    final @Nullable String errorMessage =
+        Optional.ofNullable(allowLoginRules.checkLoginIsAllowed(response, user))
+            .orElseGet(() -> authenticateUser(response, user));
+
     if (errorMessage == null || errorMessage.equals(ServerQuarantineConversation.CHANGE_PASSWORD)) {
       database.getAccessLogDao().insert(user, getUserTypeFor(response));
       return errorMessage;
@@ -97,77 +92,12 @@ public final class LobbyLoginValidator implements ILoginValidator {
   }
 
   private @Nullable String authenticateUser(final Map<String, String> response, final User user) {
-    final String clientVersionString = response.get(LobbyLoginResponseKeys.LOBBY_VERSION);
-    if (clientVersionString == null) {
-      return "No Client Version";
-    }
-    final Version clientVersion = new Version(clientVersionString);
-    if (!clientVersion.equals(LobbyConstants.LOBBY_VERSION)) {
-      return "Wrong version, we require "
-          + LobbyConstants.LOBBY_VERSION.toString()
-          + " but trying to log in with "
-          + clientVersionString;
-    }
-    if (database.getBadWordDao().containsBadWord(user.getUsername())) {
-      return ErrorMessages.THAT_IS_NOT_A_NICE_NAME;
-    }
-    if (!MacFinder.isValidHashedMacAddress(user.getHashedMacAddress())) {
-      // Must have been tampered with
-      return ErrorMessages.INVALID_MAC;
-    }
-    final Optional<Timestamp> banExpiry =
-        database.getBannedMacDao().isBanned(user.getInetAddress(), user.getHashedMacAddress());
-
-    if (banExpiry.isPresent() && banExpiry.get().toInstant().isAfter(Instant.now())) {
-      return ErrorMessages.YOU_HAVE_BEEN_BANNED + " " + getBanDurationBreakdown(banExpiry.get());
-    }
-    // test for username ban after testing normal bans, because if it is only a username ban then
-    // the user should know
-    // they can change their name
-    final boolean usernameBanned =
-        database.getUsernameBlacklistDao().isUsernameBanned(user.getUsername());
-    if (usernameBanned) {
-      return ErrorMessages.USERNAME_HAS_BEEN_BANNED;
-    }
-
-    if (!response.containsKey(LobbyLoginResponseKeys.ANONYMOUS_LOGIN)
-        && !response.containsKey(LobbyLoginResponseKeys.RSA_ENCRYPTED_PASSWORD)) {
-      return "Invalid client request";
-    }
-
     if (response.containsKey(LobbyLoginResponseKeys.ANONYMOUS_LOGIN)) {
       return authenticateAnonymousUser(response, user);
     } else if (response.containsKey(LobbyLoginResponseKeys.REGISTER_NEW_USER)) {
       return createUser(response, user);
     }
     return authenticateRegisteredUser(response, user);
-  }
-
-  private static String getBanDurationBreakdown(final Timestamp stamp) {
-    final long millis = stamp.getTime() - System.currentTimeMillis();
-    if (millis < 0) {
-      return "Ban time left: 1 Minute";
-    }
-    long seconds = Math.max(1, TimeUnit.MILLISECONDS.toSeconds(millis));
-    final int minutesInSeconds = 60;
-    final int hoursInSeconds = 60 * 60;
-    final int daysInSeconds = 60 * 60 * 24;
-    final long days = seconds / daysInSeconds;
-    seconds -= days * daysInSeconds;
-    final long hours = seconds / hoursInSeconds;
-    seconds -= hours * hoursInSeconds;
-    final long minutes = Math.max(1, seconds / minutesInSeconds);
-
-    final StringBuilder sb = new StringBuilder(64);
-    sb.append("Ban time left: ");
-    if (days > 0) {
-      sb.append(days).append(" Days ");
-    }
-    if (hours > 0) {
-      sb.append(hours).append(" Hours ");
-    }
-    sb.append(minutes).append(" Minutes");
-    return sb.toString();
   }
 
   private @Nullable String authenticateRegisteredUser(
@@ -216,26 +146,13 @@ public final class LobbyLoginValidator implements ILoginValidator {
   }
 
   private @Nullable String createUser(final Map<String, String> response, final User user) {
+    final String errMsg = allowCreateUserRules.allowCreateUser(response, user);
+    if (errMsg != null) {
+      return errMsg;
+    }
+
     final String username = user.getUsername();
     final String email = response.get(LobbyLoginResponseKeys.EMAIL);
-    if (email == null || email.trim().isEmpty()) {
-      return "Must provide an email address";
-    }
-
-    if (email.trim().contains(" ")) {
-      return "Email address may not contain spaces";
-    }
-
-    final String validationMessage =
-        Optional.ofNullable(PlayerNameValidation.validate(username))
-            .orElseGet(() -> PlayerEmailValidation.validate(email));
-    if (validationMessage != null) {
-      return validationMessage;
-    }
-
-    if (database.getUserDao().doesUserExist(username)) {
-      return "That user name has already been taken";
-    }
 
     return rsaAuthenticator.decryptPasswordForAction(
         response,
